fix(black-market): address PR #586 CI failures and CodeRabbit review
CI failures resolved:
- fmt: ran cargo fmt --all across all BM files
- clippy: removed both #[allow(clippy::too_many_arguments)] (bid.rs 8 args,
  create.rs 10 args — both under the 14-arg threshold); fixed len_zero lint
  in tests/search.rs (!is_empty instead of > 0)
- live-DB: populate connected map in make_test_state so send_to_witness_reliable
  delivers packets to TestTransport; drop invalid encrypted-packet byte-scan
  (method 92 uses Mercury extended encoding, body is encrypted — not scannable
  as raw u16 LE); assert transport delivery + DB-side invariants instead

CodeRabbit inline comments addressed:
- sweep.rs: fix stale-snapshot race — re-read full AuctionRow under FOR UPDATE
  lock so current_bid/current_bidder reflect the live state, not a pre-lock
  snapshot that a concurrent bidder could have changed
- cell_methods/black_market/mod.rs: log warn on base channel closed instead of
  silent let _ = tx.send(...); reason="base_channel_closed" on all 4 arms
- cell/cell_methods/black_market/tests.rs: drop per-field "what" inline comments
- db/sgw/BlackMarket/Tables/sgw_auction_bid.sql: amount integer -> bigint
- db/sgw/_foreign_keys.sql: seller ON DELETE CASCADE -> RESTRICT (escrow-return
  handler not yet wired); current_bidder ON DELETE SET NULL -> RESTRICT (would
  leave phantom bids current_bid>0, current_bidder IS NULL)
- .claude/agent-memory/.../project_ignore_enforcement.md: fix MD032 (blank line
  before list)

CodeRabbit out-of-diff comments addressed:
- cell_to_base.rs: condense doc comments from 730 -> 697 lines (under 700 cap)
- wire.rs push_auction_item: fresh auctions use starting_price.max(1) as
  nextMinBidPrice instead of next_min_bid(0)=1; add regression guard
  fresh_auction_next_min_bid_price_equals_starting_price

Invariants verified: #4 ghost-listing (RESTRICT FK prevents orphans until
Phase 2 cascade handler lands), #7 item-lock (no change to lock/unlock paths).

+++ db/sgw/BlackMarket/Tables/sgw_auction_bid.sql
@@ -10,7 +10,7 @@ CREATE TABLE sgw_auction_bid (
     bid_id integer NOT NULL,
     sequence_id integer NOT NULL,
     bidder_id integer NOT NULL,
-    amount integer NOT NULL,
+    amount bigint NOT NULL,
     created_at integer NOT NULL
 );
